project/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
django-crispy-forms==1.11.1
pillow==7.0.0